Cleaned up the scripts.
env_file is required to pass variables to the container (not Dockerfile), making it a variable so it's configurable in the env file.
Passing env_file param is required to compose so it itself has access to variables like Google project id.
Since container already has access to variables, easier to run a script and make passing API_HOST part of CMD rather than building container,
 plus hosting web can have different URL, will get handy later.
Expose commands seems to not be doing anything in local env, tbc